+++ OpenIddictExternalAuthentication/OpenIdAuthorizationControllerBase.cs
@@ -49,7 +49,6 @@ public abstract class OpenIdAuthorizationControllerBase<TUser, TKey> : Controlle
     /// </summary>
     protected readonly UserManager<TUser> _userManager;
 
-    private readonly IOpenIddictClientConfigurationProvider _clientConfigurationProvider;
     private readonly ILogger<OpenIdAuthorizationControllerBase<TUser, TKey>> _logger;
 
     /// <summary>
@@ -63,13 +62,11 @@ public abstract class OpenIdAuthorizationControllerBase<TUser, TKey> : Controlle
     protected OpenIdAuthorizationControllerBase(
         SignInManager<TUser> signInManager,
         UserManager<TUser> userManager,
-        IOpenIddictClientConfigurationProvider clientConfigurationProvider,
         ILogger<OpenIdAuthorizationControllerBase<TUser, TKey>> logger
     )
     {
         _signInManager = signInManager;
         _userManager = userManager;
-        _clientConfigurationProvider = clientConfigurationProvider;
         _logger = logger;
     }
 
@@ -333,7 +330,7 @@ public abstract class OpenIdAuthorizationControllerBase<TUser, TKey> : Controlle
     protected virtual async Task UpdateUser(TUser user, ExternalLoginInfo externalLoginInfo) { }
 
     /// <summary>
-    /// Tries to authorize the user user built-in method without using any specific provider.
+    /// Tries to authorize the user built-in method without using any specific provider.
     /// Usually this means showing an authentication form.
     /// </summary>
     protected virtual async Task<IActionResult> AuthorizeUsingDefaultSettings(
@@ -539,28 +536,6 @@ public abstract class OpenIdAuthorizationControllerBase<TUser, TKey> : Controlle
         }
 
         var principal = await _signInManager.CreateUserPrincipalAsync(user);
-        if (
-            !string.IsNullOrEmpty(request.ClientId)
-            && _clientConfigurationProvider.TryGetConfiguration(
-                request.ClientId,
-                out var configuration
-            )
-        )
-        {
-            if (configuration.RefreshTokenLifetime != null)
-            {
-                principal.SetRefreshTokenLifetime(
-                    TimeSpan.FromSeconds(configuration.RefreshTokenLifetime.Value)
-                );
-            }
-
-            if (configuration.AccessTokenLifetime != null)
-            {
-                principal.SetAccessTokenLifetime(
-                    TimeSpan.FromSeconds(configuration.AccessTokenLifetime.Value)
-                );
-            }
-        }
 
         await AddClaims(principal, user, request);